Make DiscreteChar storage immutable
In our project a lot of memory is spent on keeping distinct DiscreteChars which are just point-massess.
We want to share these to save some memory. Distributions have mutable interface and sharing
mutable objects is a bad idea. That is why we decided extract DiscreteChar into immutable struct
which can be safely shared among DiscreteChars.

List of changes:
- DiscreteChar state is moved into DiscreteChar.Storage struct. Which is immutable
- DiscreteChar.StorageBuilder class was introduced for stateful Storage construcation
- DiscreteChar.StorageCache class was introduced to hold single Storage for most common DiscreteChars: point massess and character classes