Remove DiscreteChar.ProbabilityOutsideRanges (#188)
A lot of code existed that had to treat `ProbabilityOutsideRanges` in a special way.
Now in cases where `ProbabilityOutsideRanges` was non-zero missing ranges are added to
cove all char domain.

`ProbabilityOutsideRanges` had 1 useful property: it needed 2 times less ranges for representing distributions distributions covering whole domain

This property is never used in real code. (And even if it did, it would be in a very few places), so reducing code complexity trumps small performance/space gain.