Implement concatenation operation for automata and transudcers (#223)
`AppendInPlace()` is not inplace anymore, it creates a copy of automata under the hood.
This makes calling it repeatedly quite expensive, because runtime will grow quadratically
with the number of calls. This method should be removed in future.

As a replacement a new method is introduced - `Concatenate()` which concatenates multiple
automata/transducers at once. It is significantly faster.